task

+++ src/tsort.h
@@ -1,10 +1,8 @@
 #ifndef TSORT_H
 #define TSORT_H
 
-#include <stdint.h>
-
 struct imm_state;
 
-int tsort(struct imm_state **states, uint16_t nstates, uint16_t ntrans);
+int tsort(struct imm_state **states, unsigned nstates, unsigned ntrans);
 
 #endif
